Move theme to libs
