+++ src/App.js
@@ -10,7 +10,7 @@ import Tour from "reactour";
 import { GrLogout } from "react-icons/gr";
 import * as Sentry from "@sentry/react";
 import sortby from "lodash.sortby";
-import { Slide, Dialog, Box, ThemeProvider, createTheme } from "@mui/material";
+import { Slide, Dialog, Box, ThemeProvider } from "@mui/material";
 import { strings } from "./libs/strings";
 import BottomNav from "./components/BottomNav";
 import sanity from "./libs/sanity";
@@ -26,6 +26,7 @@ import { getUser } from "./state/profile";
 import { errorToast } from "./libs/toasts";
 import Routes from "./Routes";
 import question from "./assets/help.png";
+import theme from "./libs/theme";
 
 const Transition = React.forwardRef(function Transition(props, ref) {
   return <Slide direction="up" ref={ref} {...props} />;
@@ -38,37 +39,6 @@ const Transition = React.forwardRef(function Transition(props, ref) {
  * @TODO Internalization Rerender - Issue #6
  */
 
-const theme = createTheme({
-  palette: {
-    mode: "dark",
-    primary: {
-      main: "#DE4665",
-      dark: "#DE4665",
-    },
-    secondary: {
-      main: "#DE4665",
-      dark: "#DE4665",
-    },
-    background: {
-      paper: "rgb(34, 34, 34)",
-      default: "rgb(0, 0, 0)",
-    },
-    text: {
-      secondary: "rgb(255, 255, 255)",
-      primary: "rgb(242, 243, 243)",
-    },
-  },
-  typography: {
-    fontSize: 14,
-    h1: {
-      fontSize: "3rem",
-      fontWeight: 300,
-    },
-    button: {
-      fontWeight: 800,
-    },
-  },
-});
 class App extends Component {
   constructor(props) {
     super(props);
